\name{Celsius}
\alias{Celsius}
\title{
Fahrenheit to Celsius
}
\description{
Accessory function to transform Fahrenheit into Celsius.
}
\usage{
Celsius(Fahrenheit)
}

\arguments{
  \item{Fahrenheit}{
Temperature in Fahrenheit.
}
}
\value{
Temperature in Celsius 
}

\author{
Ricardo Ochoa, Ilse Avalos 
}

\seealso{
\code{\link{Fahrenheit}}
}
\examples{
Celsius(64)

}
\keyword{ Celsius}
