\name{Fahrenheit}
\alias{Fahrenheit}
\title{
Celsius to Fahrenheit
}
\description{
Accesory function to transform celsius into fahrenheit
}
\usage{
Fahrenheit(celsius)
}
\arguments{
  \item{celsius}{
Temperature in celsius
}
}
\value{
Temperature in fahrenheit 
}
\author{
Ricardo Ochoa, Ilse Avalos}
\seealso{
\code{\link{Celsius}}
}
\examples{
Fahrenheit(33)
}
\keyword{Fahrenheit}
